Refactor ask command settings into separate files
