+++ Program.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using System.CommandLine;
 using System.Reflection;
+using Spectre.Console;
 
 namespace AskLlm;
